pyproject.toml
[tool.poetry]
name = "tracking-delivery"
version = "0.1.0"
description = ""
authors = ["hsw0905 <hsw0905@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
pydantic-settings = "^2.1.0"
pydantic = "^2.5.2"
inject = "^5.2.0"
flask-sqlalchemy = "^3.1.1"
requests = "^2.31.0"
celery = "^5.3.6"
flower = "^2.0.1"
redis = "^5.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
factory-boy = "^3.3.0"
pytest-factoryboy = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
